Cart
User can add items to cart. Cart item data is neither stored in session variable nor in cookie but rather in database for users and guests alike. Customer ID is stored as user_id if logged in or as session_token if guest. User can view cart and can remove items.